+++ pytket/conanfile.py
@@ -32,7 +32,7 @@ class pytketRecipe(ConanFile):
         cmake.install()
 
     def requirements(self):
-        self.requires("tket/1.2.67@tket/stable")
+        self.requires("tket/1.2.73@tket/stable")
         self.requires("tklog/0.3.3@tket/stable")
         self.requires("tkrng/0.3.3@tket/stable")
         self.requires("tkassert/0.3.4@tket/stable")
@@ -41,5 +41,5 @@ class pytketRecipe(ConanFile):
         self.requires("symengine/0.11.1")
         self.requires("gmp/6.2.1")
         self.requires("pybind11/2.11.1")
-        self.requires("nlohmann_json/3.11.2")
+        self.requires("nlohmann_json/3.11.3")
         self.requires("pybind11_json/0.2.13")