+++ tket/src/Placement/NoiseAwarePlacement.cpp
@@ -123,6 +123,35 @@ std::vector<std::map<Qubit, Node>> NoiseAwarePlacement::get_all_placement_maps(
     const Circuit& circ_, unsigned matches) const {
   std::vector<WeightedEdge> weighted_pattern_edges =
       this->default_pattern_weighting(circ_);
+
+  if (weighted_pattern_edges.empty()) {
+    // => there are no two-qubit gates in the input circuit
+    // in this case, as this method is "noise-aware", assign
+    // qubits in the circuit to the lowest-error architecture nodes
+    std::vector<std::pair<gate_error_t, Node>> all_node_errors;
+    for (const Node& node : this->architecture_.nodes()) {
+      all_node_errors.push_back(
+          {this->characterisation_.get_error(node), node});
+    }
+    // make sure all_node_errors goes from best->worst error rates
+    // N.B. `get_error` returns "0", i.e. no error, if the `Node` passed
+    // is not held in the characterisation
+    std::sort(
+        all_node_errors.begin(), all_node_errors.end(),
+        [](const auto& lhs, const auto& rhs) {
+          // Compare based on the float values in descending order
+          return lhs.first < rhs.first;
+        });
+    std::vector<Qubit> circ_qubits = circ_.all_qubits();
+    TKET_ASSERT(all_node_errors.size() >= circ_qubits.size());
+    std::map<Qubit, Node> placement_map;
+    // assign circuit qubits to Node with best error rate
+    for (std::size_t i = 0; i < circ_qubits.size(); i++) {
+      placement_map.insert({circ_qubits[i], all_node_errors[i].second});
+    }
+    return {placement_map};
+  }
+
   std::vector<boost::bimap<Qubit, Node>> placement_maps =
       this->get_all_weighted_subgraph_monomorphisms(
           circ_, weighted_pattern_edges, true);